requirements.txt
numpy >= 1.20.0
torch >= 1.9.0
torchvision
matplotlib
numpy
scipy
pillow
urllib3
scikit-image